fix(lint): resolve all golangci-lint issues
- goconst: use existing LabelValueTrue constant for "true" string
- gocyclo: extract deployMonitoring to reduce deployWorkloads complexity
- gofmt: reformat configmaps.go and ocp.go
- unconvert: remove unnecessary int32 casts in networkpolicies.go
- unparam: inline constant params in kagenti_label_test.go helpers
- unused: remove unused kagentiAgentsCRD constant

+++ internal/resources/networkpolicies.go
@@ -531,10 +531,10 @@ func apifrontendNetworkPolicy(kn *kubernautv1alpha1.Kubernaut, sidecar KagentiSi
 		metricsPort = 9092
 	}
 	if kn.Spec.APIFrontend.HealthPort != nil {
-		healthPort = int32(*kn.Spec.APIFrontend.HealthPort)
+		healthPort = *kn.Spec.APIFrontend.HealthPort
 	}
 	if kn.Spec.APIFrontend.MetricsPort != nil {
-		metricsPort = int32(*kn.Spec.APIFrontend.MetricsPort)
+		metricsPort = *kn.Spec.APIFrontend.MetricsPort
 	}
 	pHealth := intstr.FromInt32(healthPort)
 	pMetrics := intstr.FromInt32(metricsPort)
